dashboards/requirements.txt
flask
kafka-python
matplotlib
flask-cors
